typo fixes
Doctor availabilities are weird as of now:
So with the scope, you get an issue where the end time given must exceed the end time in db of an availability for it to count.
ie: entry 8am to 10am, query 8am to 9am should be valid but since it isn't covered by the given dates its not accepted.